Optimize final part of allocation hot path
In the seastar allocator C++ allocation path there are two checks
for the returned pointer being null. One to call the
on_allocation_failure function, and another (C++ specific) one to
throw std::bad_alloc if the allocation has failed.

On clang the optimizer wasn't able to eliminate the redundant check so
we hold its hand a bit, moving a stats allocation out of the critical
region and hinting a branch direction.

This saves a couple of instructions on the hot path.