Remove null pointer check on free hot path
In the various free hot paths we check if the pointer is null and
return immediately if so. This ads a compare and branch to the hot
path (well, every path) for an unusual case. Remove this and let
the logic in the free path handle it more slowly but benefiting every
non-null free.