file: construct directory_entry with aggregated ctor
despite that `directory_entry` can be default constructed, it does not
have a contructor which accepts two parameters. so, in order to
construct an instance of it, we either need to use the aggregated
contructor, or the list-initialization.

in this change, in order to appease the compiler which does not accept
the form of construct, let's use the aggregated constructor.

this should address the FTBFS with Clang-15 + C++20 after the commit of
95a7ff99:

```
/home/circleci/project/src/core/file.cc:417:29: error: no matching constructor for initialization of 'seastar::directory_entry'
            directory_entry ret(std::move(name), type);
                            ^   ~~~~~~~~~~~~~~~~~~~~~
/home/circleci/project/include/seastar/core/file.hh:58:8: note: candidate constructor (the implicit copy constructor) not viable: requires 1 argument, but 2 were provided
struct directory_entry {
       ^
/home/circleci/project/include/seastar/core/file.hh:58:8: note: candidate constructor (the implicit move constructor) not viable: requires 1 argument, but 2 were provided
/home/circleci/project/include/seastar/core/file.hh:58:8: note: candidate constructor (the implicit default constructor) not viable: requires 0 arguments, but 2 were provided
1 error generated.
```

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>

Closes scylladb/seastar#1909