fair_queue: Do not pop unplugged class immediately
A fair queue's priority class have two bits -- queued and plugged.

Queued means that the class is on the priority queue and dispatching
loop will eventually pick it up from there. When false this means
that the class is not on the handles list. The bit protects the class
from being enqueued more than once.

Dequeueing (called popping) has a trickery -- it can only happen on the
class that's on top of the dispatch queue. That's because the dispatch
queue queue is std::priority_queue which doesn't have random access
eraser call, but only have the .pop() one.

Plugged bit is controlled by the bandwidth throttler. When the class
runs out of its throttling tokens, it's unplugged. When it replenishes
the necessary amount of tokens backm, it's plugged back. Currently,
unplugging the class also dequeues it from the dispatch list.

The aforementioned dequeuing limitation works with unplugging most of
the time, because unplug can only happen from inside the dispatch loop
and and the dispatch loop first picks the class from the top of the
queue. So unplugged class is automagically the top one.

However, if there's duplex mode turned on, then each class is present in
two fair queues (called streams) and unplugging a class means unplugging
it from both fair queues. And even if the priority class being
dispatched is indeed the top one, its couterpart from another stream is
not, and popping it while unplugging is wrong and leads to broken pop --
the correct class is maked as unplugged and dequeued, but it's _not_
removed from the dispatch list, instead some other class is pooped from
it. Later dispatch loop would find this other class in the dispatch
queue makred unplugged and dequeued and would crash on assertion.

The fix is to pop classes lazily. It means that unplugged classes are
just marked as such and that's it. Later, when the dispatch loop picks
up the unplugged class is will pop it instantly and proceed.

fixes: #1899

Tested with 10 minutes of io-tester with the reproducer config attached
to the aforementioned issue.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>