build: do not error when running into GCC BZ-1017852
if the compiler impacted by
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=107852 is used to compile
Seastar with fmtlib v10, the tree would fail to compile, because we
consider `-Warray-bounds` warnings as error.

to build with compiler (typically GCC 13.1) which suffers from this
bug. let's stop considering these false alarms as error. a minimal
reproducer is introduced for testing this compiler bug.

Fixes #1716
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>