shared_future: make available() immediate after set_value()
A shared future takes a regular future on construction and resolves
after the original future resolves. This is implemented by attaching
a `then_wrapped` continuation on the original future which, when run,
forwards the result of the original future to the currently registered
waiters.

The original future is kept in the `_original_future` field. The
continuation is attached on the first `shared_future::get_future()`
call - which causes the `_original_future` to become invalid - and
it is reassigned again at the end of the continuation.

This implementation detail has an effect on
`shared_future::available()`, which is implemented via
`_original_future.available()`. When the promise associated with the
original future is set, the continuation is scheduled for execution
but does not run immediately. Because of that, `_original_future`
is not updated immediately after `promise::set_value`
and `shared_future::available()` keeps returning false until
the continuation runs.

This is especially confusing for `shared_promise`, which are
essentially a `promise` + `shared_future` pair and have both
`set_value` and `available` methods: setting the shared promise does
not cause it to become available.

Fix this by implementing the `task` interface for the shared state.
Instead of creating a continuation via `.then_wrapped`, the shared state
is attached directly as a task to the future. There is no longer need to
move the `_original_future` back and forth so the `available()` method
now works correctly.

Fixes: scylladb/seastar#1926