net: Fix single cpu packet reassembly fault
In dpdk.cc:1515 when running program utilizing Seastar with smp::count == 1,
rss hashing is NOT enabled, but when a fragmented packet is reassembled in
ip.cc:195, no check is done if smp::count == 1, so the following line
is called to identify cpu to forward packet to:

cpu_id = _netif->hash2cpu(toeplitz_hash(_netif->rss_key(), hash_data));

Since running with only a single cpu and hashing not configured,
the call to toeplitz_hash faults.

Issue resolved by checking to see if smp::count == 1 after defragmentation
and if so, bypassing hash calls and just calling l4->received()
directly.