test_runner: do not throw when seastar.app fails to start
we feed seastar::app with the command line options passed in the
command line as they are, and seastar::app chokes at seeing unknown
options.

in `test_runner::start_thread()`, we throw in this case, but
`test_runner::run_sync()` expects that the worker thread to
start the reactor and then to serve the tasks, without throwing
any exception. while, if an exception is thrown in the worker
thread, the worker thread exits. the parent thread is still
waiting and expecting the worker thread to signal it for serving,
that's why the test executable hangs at seeing unknown arguments.

in this change, instead of throwing at unknown arguments or when
unable to start the reactor, we just report it back to the caller,
so that it can recored the failure as `_exit_code`, and skip this
test and all following tests. so `alloc_test  -- c2` does not hang
anymore.

Fixes #673
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>