coroutine: generator: reset _wait_for_next_value when set_value()
instead of using a sequence of

1. _wait_for_next_value->set_value();
2. _wait_for_next_value = {};

let's just reset it before set_value() on it, this ensures that the
coroutine(s) waiting on this promise, to see an empty
`_wait_for_next_value`. they don't have to assert this, but it
would simplify their lives and improve the readability by making the
invariant more clear.

Fixes #1913
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>