build: let tri_state_option() accept multi build build types
this addresses a regression introduced by 0a6d88ad, which let
DEFAULT_BUILD_TYPE accept a single value instead multiple values,
and let CONDITION accept multiple values. but we do intend to
pass multiple modes as the DEFAULT_BUILD_TYPE option.

so in this change, the arguments passed to `cmake_parse_arguments`
is transposed to correct this regression.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>