build: do not return a non-false condition if the option is off
before this change, tri_state_option() sets a condition which
can be evaluted to TRUE at configure time, even if the option is
disabled. and this condition is evaluated to false by the
per-configuration generator. so, under most circumstances, this
works. but when it comes to `Seastar_SANITIZE`, it also sets a
variable `Seastar_Sanitizers_OPTIONS` as long as the `condition`
is evalutated to TRUE at the configure time. and
`Seastar_Sanitizers_OPTIONS` is in turn used in the .pc file as
part of the `seastar_libs` variable. in other words, the
sanitizer options is _always_ used in all build modes. this is
not expected.

so, in this change, we do not return a condition string which
can be evaluated to TRUE if we are using a non-multi-config
generator and the build mode is not listed in the default mode
list. the behavior when building with mult-config generator
is preserved.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>