shared_future: test for dangling tasks
The previous version of the PR was reverted in scylladb/seastar#1937 due
to a slight change in how shared_future works internally: when a
shared_future was created from a not-yet-resolved future, it would
schedule a task to run after the future gets resolved. In case when, at
the end of the program, there are some unsatisfied shared_promises
(which are essentially promise + shared_future pairs) and they get
destroyed, the promises would be broken and shared futures' internal
tasks would get scheduled for execution. However, as nobody waits for
the tasks, it could happen that the program exits before they have a
chance to run. This is harmless by itself but causes leak detectors to
complain (example: LeakSanitizer in ScyllaDB's `schema_registry_test`
test).

The new version does not have this problem and schedules the task in
cases when the master implementation does - i.e. get_future must be
called, so the task is indirectly waited on. Add a test which verifies
that the internal task is scheduled at appropriate moments.