tls: add format_as(subject_alt_name_type) overload
this change addresses the formatting with fmtlib v10, which stopped
creating the default-generated formatter even if FMT_DEPRECATED_OSTREAM
preprocess macro is defined. but it allows us to format by providing
a format_as() overload. please note this does not work with fmtlib v9,
in which, we still need either define FMT_DEPRECATED_OSTREAM or define
a proper fmt::formatter<> for this type.

Refs https://github.com/scylladb/scylladb/issues/13245
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>