file: use unbuffered generator in experimental_list_directory()
because the buggy buffered generator, as a workaround, let's use
the unbuffered generator as an intermediate workaround of #1913 .

the downside of this change is that buffered generator allows the
generator coroutine to yield the elements without being suspended,
if it is able to generate the produced elements in batch, thus can
have better performance. this design is a good fit of use cases
like readdir. we will revert this change once #1913 is fixed.

Refs #1913
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>