coroutine: generator: use std::optional for container by default
because the buffered generator suffered from #1913, and because
clang takes an templates alias as a different template, and refuses
to use the partial specialization of `generator<T, std::optional>`,
if the `Container` is an alias of `std::optional`. we cannot
define a template alias of `std::optional` and pass it to
`coroutine::experimental::generator<>` as its template parameter,
to simplify the use of generator, let's define a default template
parameter.

Refs #1913
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>