http: make_requests accept optional expected
The original API requires passing the expected status into the
make_request function. The reply_handle function will only be
called if the expected status matches. Otherwise, a
httpd::unexpected_status_error will be returned and there is
no way to handle the response body, which contains additional
information. With the httpd::unexpected_status_error,
reply->status is not enough. This was discussed in detail on #1963.

This commit patches the API in a backward incompatible way.
If the expected status is std::nullopt, the reply_handler will
handle any reply without missing anything.