io: Apply rate-factor early
Here's the sched throttling math:

  bw/bw_max + iops/iops_max <= C

The "rate_factor" is the C and it's usually set to be 1.0, but there was
a reservation for convex/concave io profiles where the C would be set to
lower or higher values to adjust the limit.

The same goal can be achieved by scaling ..._max values on the right
side of the equation, like this

  bw/(bw_max * C) + iops/(iopx_max * C) <= 1.0

This makes the fair-queue code a bit simpler and, as a nice side effect,
adds sanity check for the rate-factor not been too low.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>