sharded: access instance promise only on instance shard
Currently, the vector of instances is allocated on shard 0,
including a promise<> per instance, yet those promises
are accessed on each instance shard, which is susceptible
to #1563.

To prevent that, move the promise to `async_sharded_service`,
as it is used only with this class, and add
sharded::entry::track_deletion() that gets a future from the said
promise on the shard it was allocated on if the Service
base class is `async_sharded_service<Service>`.

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>