future: promise_base: assert_task_shard in debug mode
Added SEASTAR_DEBUG_PROMISE macro that is defined
in debug and sanitize modes.

Under that #ifdef, keep _task_shard := this_shard_id()
along with _task on the promise_base, in all cases
that a task is scheduled to run when the promise
becomes ready.

In promise_base::make_ready, assert that the shard
that sets the promise value/exception is the same
as the shard waiting for the task to run, otherwise
the waiting task will be scheduled on the wrong
shard.

Fixes scylladb/seastar#1563

Signed-off-by: Benny Halevy <bhalevy@scylladb.com>