build: install Finducontext.cmake and FindSystem-SDT.cmake
in 6d2b0a72, Finducontext.cmake was introduced to find libucontext
if necessary. and ucontext::ucontext was added as yet another linkage
of Seastar::seastar if Seastar is built in Release mode as one of its
INTERFACE_LINK_LIBRARIES. but we failed to install Finducontext.cmake
along with the Seastar's module config file. so, if user configure
the building system for an application with a Seastar installed into
the system, cmake would fail like:

```
CMake Warning (dev) at /home/ubuntu/.local/lib/cmake/Seastar/SeastarTargets.cmake:61 (set_target_properties):
  Policy CMP0028 is not set: Double colon in target name means ALIAS or
  IMPORTED target.  Run "cmake --help-policy CMP0028" for policy details.
  Use the cmake_policy command to set the policy and suppress this warning.

  The link interface of target "Seastar::seastar" contains:

    ucontext::ucontext

  but the target was not found.  Possible reasons include:

    * There is a typo in the target name.
    * A find_package call is missing for an IMPORTED target.
    * An ALIAS target is missing.

Call Stack (most recent call first):
  /home/ubuntu/.local/lib/cmake/Seastar/SeastarConfig.cmake:36 (include)
  CMakeLists.txt:1 (find_package)
This warning is for project developers.  Use -Wno-dev to suppress it.
```

the same applies to SystemTap-SDT.
in this change,

* ucontext is checked using SeastarDependencies.cmake, so we can check
  for it when building Seastar and when building Seastar applications.
* Finducontext.cmake is installed.
* apply the same change to SystemTap-SDT

Refs 6d2b0a72
Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>