install-dependencies.sh: add libudev-dev and systemd-devel
Seastar supports both static and dynamic linkage. and we have
command line example in README.md to statically link against
Seastar. this means, all of its linkages are statically linked.

if hwloc is statically linked. it would bring into libudev, See

```
$ pkg-config --libs --cflags --static hwloc
-lhwloc -lm -ludev -lpthread
```
on Debian and its derivative distros, libudev.so is packaged in
libudev-dev, while on Fedora and its derivative distros, libudev.so
in systemd-devel.

in this change, these packages are included in install-dependencies.sh,
so the building system can have access to them when linking Seastar
statically. please note, these packages are not dependencies of
libhwloc-dev or hwloc-devel. Because, in general, distro packacage
guidelines encourage maintainers to package shared library if possible,
and reuse the shared libraries redistributed by the distro whenever
appropriate.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>