memory: do not set_reclaim_hook if cpu_mem_ptr is not set
in the release build, we use Seastar allocator, but the Seastar application
can still configure `smp_opts.memory_allocator` with `memory_allocator::standard`.
in that case, the `memory::configure()` is not called, hence `cpu_mem_ptr` is not
set.

but in the constructor of `reactor`, it is supposed to setup per-reactor
facilities, notably the ones backed by TLS variables, including the
memory subsystem. but we are not able to access the `smp_opts` anymore
in `reactor::reactor()`. because the memory subsystem is configured in
`smp::configure()`, where is `smp_opts` passed in.

so, in this change

* instead of conditionally calling
  `memory::set_reclaim_hook()` in `reactor::reactor()` based on the
  value of `smp_opts.memory_allocator`, we check for null in
  `memory::set_reclaim_hook()`. fortunately, this is not a hot path,
  we only call `memory::set_reclaim_hook()` when starting the reactor,
  so the extra check does not hurt the performance.
* add a test for exercising template_app, as the smp options is
  a public interface for customizing the behavior of app_template.
  because boot.test considers it an error if a test suite does not
  contains any test, and fails the test with:
  ```
  Test setup error: test tree is empty
  ```
  we cannot disable the test in app-template_test.cc using
  `boost::unit_test::enable_if<>`. neither can we add more
  test cases in this test, because seastar cannot be initialized
  multiple times in a single process. so we have to disable
  this test in the CMake.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>