exception_hacks: do not initialize phdrs_cache multiple times
`init_phdr_cache()` is called when in `smp::configure()`. but
if we teardown an app_template and start a new one, and both
seastar applications has
`smp_opts.enable_glibc_exception_scaling_workaround` set,
the `dl_phdr_info` vector would be initialized twice. this is
a false problem in production, as we don't launch seastar app multiple
times in a single process. but this is handy if we want to test,
for instance, app_template in a single test process, where
we need to start multiple app_template instances.

so, in this change, we just don't set `phdrs_cache` again if
it is already set.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>