test: add more tests for app_template
this test exercises the app_template's functionalities. more importantly,
it verifies that app_template is able to run multiple times in a singple
process. this feature is important for running multiple tests exercising
app_template.

since the new test always contains one or more test cases, even if dpdk
is enabled. we can now disable the test which tests the standard memory
allocator in the test without using CMake.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>