treewide: s/format/fmt::format/ when appropriate
this change replace some `format()` calls with `fmt::format()`.

before this change, we use `format()` when formatting to an sstring
and when formatting to an std::string.

the upside of this design is that we don't need to differentiate
the calls where an sstring is expected from those where an std::string
is expected.

but the downside is that we create an `std::string` by deep copy
the content of sstring after formatting the parameters and then drop
the `sstring` on the floor in the second use case, where only an
std::string is expected.

another potential problem is that there are actually three
`format()` functions. they are from three different namespaces:
`seastar`, `std` and `fmt` in C++20. currently, the `seastar::format()`
one wins in the function loopup. because the other two are
templated format string, while `seastar::format()` accepts a plain
`const char*`, and we always open the `seastar` namespace in this
project.

and in future, once we switch over to the compile-time format check,
the function signature of `seastar::format()` will be changed from
`format(const char* fmt, A&&... a)` to
`format(fmt::format_string<A...> fmt, A&&... a)`, the compiler would
have trouble when performing the function lookup because of the ambiguity.

so, to address the second problem we have to change all
call sites of this function to one of these three format functions.

but before that, just to prepare for the change, let's change some of
them to `fmt::format()` to address both the first problem and the
second one.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>