prometheus.cc: Clean the pick representation code
Prometheus supports two representations: text and protobuf. Text is the
default representation.
The decision logic is: it is text format unless someone asks explicitly for protobuf.

This patch makes it more evident in the code by making two changes:
1. Rename is_accept_text to is_accept_protobuf.
2. Remove negations by renaming is_text_format to is_protobuf_format.

Those changes do not affect the overall logic, but the code is now more
straightforward.

Signed-off-by: Amnon Heiman <amnon@scylladb.com>