README-DPDK.md: point user to DPDK's quick start guide
DPDK has renamed `setup.sh` to `dpdk-setup.sh`, and later dropped it
in v20.11. while in latest Seastar we point the dpdk submodule to
a fork based on v23.07 .

also, the igb_uio kernel module was moved out of dpdk. and vfio-pci
is the recommended kernel driver to work with the userspace PMD drivers.
see https://doc.dpdk.org/guides/linux_gsg/linux_drivers.html

and DPDK's quick start guide also covers the step to configure
hugepage.

so, let's update the document and point user to DPDK's document, which
is more updated.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>