metrics_tester - A demo app to test metrics
The challenge with testing the metrics protobuf protocol in general and
testing the experimental native-histogram specifically is that it's not
enough to produce a valid protobuf message.  We need to validate that
Prometheus agrees with what is reported.

The metrics_tester is a base for such a test.
The idea is to report a small subset of metrics with known values and
allow accepting protobuf requests.

To test it, you will need to run a Prometheus server with version 2.4 or
higher and set --enable-feature=native-histograms. You could then query
the Prometheus server and see if you get the correct values.

At this phase, it is done manually; full automation will be added in the
future.

The app currently registers only two metrics, a gauge, and a histogram.
It is here for basic validation only.

Signed-off-by: Amnon Heiman <amnon@scylladb.com>