shared_future: shared_state::run_and_dispose(): release reserve of _peers
run_and_dispose() clears _peers but if it was non empty at some point,
it might still hold on to some reserve memory after it was cleared.
After run_and_dispose() was run, it won't ever contain members again, so
this is just a waste, kept around until the shard_state is finally
destroyed. Since _peers is a chunked_fifo under the hood, this
memory amount might be significant too.
Prevent this waste by resetting _peers to default state at the end of
said method.

Refs: scylladb/scylladb#16493

Closes #2033