metrics_tester: Add a test app test_metrics.py
The test_metrics.py is a utility that validates that registered metrics
are presented as expected.
Its primary purpose is Native histogram validation and Prometheus
proto-buf, but it also validates text format to make sure that any
changes in Protobuf representation will not affect the text
representation.

The validation assumes a metrics_tester app running and a Prometheus
server running with protobuf enable.
test_metrics.py will read the text representation directly from the
metrics_tester and will compare the results to the config fie.
It will then use the Prometheus query API to validate that it gets the
same values.

A note about histogram validation: there are minor differences between
Prometheus native_histogram, and seastar
approximate_exponential_histogram.

In Seastar, the buckets are pseudo-exponential, while Prometheus uses
exponential buckets.
Also, in Seastar (as in the text representation), the buckets are
left-open, while in native histograms, they are right-open.
test_metrics.py takes that into account.