net/posix: Track ap-server ports conflict
If creating two listening posix sockets with conflicting ports on
non-main thread, the listen() call succeeds, but subsequent
server_socket::accept() steps on assertion.

The thing is that posix_ap_server_socket_impl() accept() may try to
emplace a promise into thread-local sockets map which is used to wait
for the main thread to accept() the connection for real and then send it
to this map. The map uses server socket as a key in the map, and when
emplaced, the code asserts that insertion did took place, not the
existing object was found. This is because listening can only be done
once, if there's no listened, main thread sends ready sockets via
another thread-local container.

When two ap server sockets bound to the same port call accept() only one
of them them manages to _insert_ the promise -- the other one will find
that the promise is already there and assertion fires.

The fix is in tracking ports for ap stack and throwing EADDRINUSE in
case of conflict early.

fixes: #2036

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>

Closes scylladb/seastar#2040