metrics_api.hh: Remove duplication of metric family metadata
Metric family metadata consists of information relevant to all metrics
with the same name (but with different labels).  For example, name,
description, and aggregate labels.

The metric family metadata is found in two data structures: the
_value_map that holds all registered metrics and _metadata that contains
only reported metrics.

This patch makes the _metadata hold a reference to the data in the
_value_map instead of copying it, which serves two purposes:
1. Minor performance gain (memory and computation).
2. It will now be possible for a change in _value_map to be reflected in
   _metadata.

Signed-off-by: Amnon Heiman <amnon@scylladb.com>