net: add missing include
A recent change to net/posix-stack.hh added a use of std::unordered_set
but forgot to include the header file <unordered_set>. This was fine
on some versions of the standard library, but ended up in a missing
the include in other versions. So let's add the missing include.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>

Closes scylladb/seastar#2043