rpc: extend the use_gate until request processing is finished
Before the patch we released the use_gate in
server::connection::process when the future,
returned by handler func is complete. This
doesn't mean though that the request processing
is finished, it continues in the background and
can cause segfaults during node shutdown when
unregister_handler is called.

To resolve this we extend the use_gate until all
request related background activity is finished.

Closes scylladb/seastar#2044