Update supported C++ standards to C++23 and C++20 (dropping C++17)
C++23 has passed all the ballots [1]. We can therefore support it now
formally. This means that C++17 is no longer supported, and C++20 is
supported until C++26 is released.

The primary benefit of C++23 support is that we can now use coroutines
in core Seastar code, as all supported C++ versions have them.

[1] https://lists.isocpp.org/std-proposals/2024/01/8880.php