build: prevent gcc -Wstringop-overflow workaround from affecting clang
We check for a gcc-specific bug, but because we inject warning flags,
the test fails on newer versions of clang (since it doesn't like the
warning flags). Due to the failure, we then inject the workaround for
clang builds, which again fails.

Fix that by only applying the workaround for gcc.

Closes scylladb/seastar#2055