future: deprecate future::get0()
In the days when futures carried tuples, it was useful to get
the first/only element. But it's now a synonym for get(), so
deprecate it.

Also, all future::get0() calls were dropped in place of future::get()
in 1504dc0e, so Seastar should be free of deprecation warnings
due to this change.

Kefu: amend the comment to highlight the difference between `get0()`
and `get()`, to help developers to switch from `get0()` to `get()`.

Closes #2188