Clarify reactor utilization metrics
The key CPU use metric for seastar is "reactor utilization". This is
the time spent on CPU [1] except that doing busy polling (i.e., when
there is no additional work to do but we continue polling for IO because
"why not" and "that's how we set things up to get disk completitions").

However, we simply call this "CPU utilization" or "CPU busy time" in the
metrics, which is confusing, since it is very different from actual
OS-reported CPU utilization because it excludes polling time. This is
a frequent source of confusion when people expect this to act like
CPU utilization or use this interchangeably with CPU utilization
reported by the OS (e.g., via node_exporter).

This change clarifies the description, calling it "Reactor non-polling
CPU utilization". Furthermore, it also clarifies the specific
definition of the utilization (gauge) variation of the metric: it is
a five second rolling average.