io_tester: implement request_type::unlink
This change introduces request_type::unlink as well
as unlink_class_data to io_tester. It also extends
the creation of operations to be executed to recognize
the new request type.

The purpose of this change is to enable io_tester to
analyze the impact of unlink operations on read and
write operations.

unlink_class_data creates a given number of files
during startup. When requests are issued it calls
unlink on the created files.

Refs: scylladb#1299
Signed-off-by: Patryk Wrobel <patryk.wrobel@scylladb.com>